♻️ Refactor: toast 관련 types는 types 폴더 내에 넣기
