+++ src/common/components/toast/Toaster/Toaster.client.tsx
@@ -1,7 +1,5 @@
 'use client';
 
-import { MAX_TOAST_COUNT } from '@common/constants/toast';
-
 import { useToastInfo } from '@common/hooks/stores/useToastStore';
 
 import Toast from '../Toast/Toast.client';
@@ -10,7 +8,7 @@ const Toaster = () => {
   const { toastList } = useToastInfo();
 
   return (
-    <div className="bottom-20pxr right-20pxr w-300pxr gap-20pxr fixed z-1 flex flex-col items-end">
+    <div className="bottom-20pxr right-20pxr w-300pxr gap-30pxr fixed z-1 flex flex-col items-end">
       {toastList.map((toast) => (
         <Toast key={toast.key} toast={toast} />
       ))}
