+++ src/common/components/toast/types/toast.types.ts
@@ -1,0 +1,12 @@
+export enum ToastType {
+  SUCCESS = 'SUCCESS',
+  ERROR = 'ERROR',
+  INFO = 'INFO',
+}
+
+export interface ToastItem {
+  key: string;
+  text: string;
+  type: ToastType;
+  duration?: number;
+}
